refactor(api): move emailVerificationSerializer
to src/identity-access-management
