+++ api/tests/identity-access-management/unit/infrastructure/serializers/jsonapi/update-email.serializer.test.js
@@ -1,7 +1,7 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/update-email-serializer.js';
-import { expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../src/identity-access-management/infrastructure/serializers/jsonapi/update-email.serializer.js';
+import { expect } from '../../../../../test-helper.js';
 
-describe('Unit | Serializer | JSONAPI | update-email-serializer', function () {
+describe('Unit | Identity Access Management | Infrastructure | Serializer | JSONAPI | update-email-serializer', function () {
   describe('#serialize()', function () {
     it('should convert user new email into JSON API data', function () {
       //given
